requirements.txt
django==4.2.11
djangorestframework==3.14.0
drf-spectacular==0.26.5
django-cors-headers==4.3.1